do not output colors if stdout is not a tty
If stdout is not a tty (e.g. a file), a user does not normally want to
have ANSI escape codes present in the output.
This commit adds a check (sys.stdout.isatty()) to verify if stdout is a
tty so it will only output color codes in this specific case.